+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v1.0.0 (2021.11.xx)
   - Add Twitter.getListTweets() for "GET /2/lists/:id/tweets"
 - List members
   - Add Twitter.getListMembers() for "GET /2/lists/:id/members"
+  - Add Twitter.getListMemberships() for "GET /2/users/:id/list_memberships"
 - Bump Kotlin 1.6.0
 
 v0.4.0 (2021.11.12)